Move all beacon_state fns into single file.
